mardi matin tchi

+++ database/migrations/2021_12_10_132230_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->id();
             $table->timestamps();
             $table->string('status');
-            $table->integer('total')->nullable();
+            $table->float('total')->nullable();
             $table->string('comments')->nullable();
             $table->string('note_admin')->nullable();
             $table->integer('user_id')->foreign()
